finally put Wapiti class and wapiti_main in separate files
