+++ tests/attack/test_mod_buster.py
@@ -7,7 +7,7 @@ import pytest
 
 from wapitiCore.net import Request
 from wapitiCore.net.crawler import AsyncCrawler
-from wapitiCore.net.crawler_configuration import CrawlerConfiguration
+from wapitiCore.net.classes import CrawlerConfiguration
 from wapitiCore.attack.mod_buster import ModuleBuster
 from wapitiCore.attack.attack import Flags
 from tests import AsyncIterator
