+++ tests/attack/test_mod_cookieflags.py
@@ -8,7 +8,7 @@ import pytest
 
 from wapitiCore.net import Request
 from wapitiCore.net.crawler import AsyncCrawler
-from wapitiCore.net.crawler_configuration import CrawlerConfiguration
+from wapitiCore.net.classes import CrawlerConfiguration
 from wapitiCore.attack.mod_cookieflags import ModuleCookieflags
