+++ tests/attack/test_mod_crlf.py
@@ -7,7 +7,7 @@ import httpx
 
 from wapitiCore.net import Request
 from wapitiCore.net.crawler import AsyncCrawler
-from wapitiCore.net.crawler_configuration import CrawlerConfiguration
+from wapitiCore.net.classes import CrawlerConfiguration
 from wapitiCore.attack.mod_crlf import ModuleCrlf
