+++ tests/attack/test_mod_exec.py
@@ -9,7 +9,7 @@ import pytest
 import respx
 import httpx
 
-from wapitiCore.net.crawler_configuration import CrawlerConfiguration
+from wapitiCore.net.classes import CrawlerConfiguration
 from wapitiCore.net import Request
 from wapitiCore.net.crawler import AsyncCrawler
 from wapitiCore.attack.mod_exec import ModuleExec
