+++ tests/attack/test_mod_htaccess.py
@@ -5,7 +5,7 @@ import httpx
 import respx
 import pytest
 
-from wapitiCore.net.crawler_configuration import CrawlerConfiguration
+from wapitiCore.net.classes import CrawlerConfiguration
 from wapitiCore.net import Request, Response
 from wapitiCore.net.crawler import AsyncCrawler
 from wapitiCore.attack.mod_htaccess import ModuleHtaccess
