+++ tests/attack/test_mod_htp.py
@@ -9,7 +9,7 @@ import respx
 
 from wapitiCore.attack.mod_htp import ModuleHtp
 from wapitiCore.net.crawler import AsyncCrawler
-from wapitiCore.net.crawler_configuration import CrawlerConfiguration
+from wapitiCore.net.classes import CrawlerConfiguration
 from wapitiCore.net import Request
