+++ tests/attack/test_mod_nikto.py
@@ -7,7 +7,7 @@ import httpx
 import respx
 import pytest
 
-from wapitiCore.net.crawler_configuration import CrawlerConfiguration
+from wapitiCore.net.classes import CrawlerConfiguration
 from wapitiCore.net import Request
 from wapitiCore.net.crawler import AsyncCrawler
 from wapitiCore.attack.mod_nikto import ModuleNikto
