+++ tests/attack/test_mod_redirect.py
@@ -9,7 +9,7 @@ import pytest
 import httpx
 import respx
 
-from wapitiCore.net.crawler_configuration import CrawlerConfiguration
+from wapitiCore.net.classes import CrawlerConfiguration
 from wapitiCore.net import Request
 from wapitiCore.net.crawler import AsyncCrawler
 from wapitiCore.attack.mod_redirect import ModuleRedirect
