+++ tests/attack/test_mod_ssl.py
@@ -9,7 +9,7 @@ from unittest.mock import AsyncMock
 
 import pytest
 
-from wapitiCore.net.crawler_configuration import CrawlerConfiguration
+from wapitiCore.net.classes import CrawlerConfiguration
 from wapitiCore.net import Request
 from wapitiCore.language.vulnerability import CRITICAL_LEVEL, HIGH_LEVEL, INFO_LEVEL
 from wapitiCore.net.crawler import AsyncCrawler
