+++ tests/attack/test_mod_wp_enum.py
@@ -7,7 +7,7 @@ import respx
 
 from wapitiCore.attack.mod_wp_enum import ModuleWpEnum
 from wapitiCore.net.crawler import AsyncCrawler
-from wapitiCore.net.crawler_configuration import CrawlerConfiguration
+from wapitiCore.net.classes import CrawlerConfiguration
 from wapitiCore.net import Request
