+++ tests/attack/test_mod_xss_basics.py
@@ -5,7 +5,7 @@ import respx
 import httpx
 import pytest
 
-from wapitiCore.net.crawler_configuration import CrawlerConfiguration
+from wapitiCore.net.classes import CrawlerConfiguration
 from wapitiCore.net import Request
 from wapitiCore.net.crawler import AsyncCrawler
 from wapitiCore.attack.mod_xss import ModuleXss
