+++ tests/cli/test_options.py
@@ -5,7 +5,8 @@ from unittest import mock
 import pytest
 from wapitiCore.attack.attack import common_modules, all_modules, passive_modules
 from wapitiCore.net import Request
-from wapitiCore.main.wapiti import Wapiti, wapiti_main
+from wapitiCore.main.wapiti import wapiti_main
+from wapitiCore.controller.wapiti import Wapiti
 
 
 @pytest.mark.asyncio
