+++ tests/cookies/test_json_cookie.py
@@ -5,7 +5,7 @@ import respx
 import httpx
 import pytest
 
-from wapitiCore.net.crawler_configuration import CrawlerConfiguration
+from wapitiCore.net.classes import CrawlerConfiguration
 from wapitiCore.net.jsoncookie import JsonCookie
 from wapitiCore.net.crawler import AsyncCrawler
 from wapitiCore.net import Request
