+++ tests/web/test_network_issues.py
@@ -7,7 +7,7 @@ import pytest
 from httpx import ReadTimeout
 
 from wapitiCore.net.crawler import AsyncCrawler
-from wapitiCore.net.crawler_configuration import CrawlerConfiguration
+from wapitiCore.net.classes import CrawlerConfiguration
 from wapitiCore.net import Request
