+++ wapitiCore/net/classes.py
